Update mice-vs-amelia-longitudinal-sim.R
Streamlined the process of generating unit-level parameters (for slop, intercept, and error) and for generating missingness at the start of a vector. Currently, I made relatively short amounts of missingness more likely by using a binomial distribution, but I would like to allow for a few units to have only one or two points observed.